requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
docopt==0.6.2
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.8.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.3
pipreqs==0.4.9
pkg-resources==0.0.0
pycrypto==2.6.1
pytz==2018.4
redis==2.10.6
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
yarg==0.1.9
